Move ThreadLocalStorage to utils module
