+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/utils/ThreadLocalDisposableHolder.kt
@@ -1,0 +1,5 @@
+package com.badoo.reaktive.utils
+
+import com.badoo.reaktive.disposable.Disposable
+
+internal open class ThreadLocalDisposableHolder<T>(initialValue: T? = null) : ThreadLocalHolder<T>(initialValue), Disposable
