+++ tools/binary-compatibility/reference-public-api/reaktive.txt
@@ -1058,14 +1058,10 @@ public final class com/badoo/reaktive/utils/HandleSourceErrorKt {
 	public static synthetic fun handleReaktiveError$default (Ljava/lang/Throwable;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
 }
 
-public final class com/badoo/reaktive/utils/ThreadLocalStorage : com/badoo/reaktive/disposable/Disposable {
+public class com/badoo/reaktive/utils/ThreadLocalStorage : com/badoo/reaktive/utils/ThreadLocalHolder, com/badoo/reaktive/disposable/Disposable {
 	public fun <init> ()V
 	public fun <init> (Ljava/lang/Object;)V
 	public synthetic fun <init> (Ljava/lang/Object;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun dispose ()V
-	public final fun get ()Ljava/lang/Object;
-	public fun isDisposed ()Z
-	public final fun set (Ljava/lang/Object;)V
 }
 
 public final class com/badoo/reaktive/utils/UncaughtErrorHandlerKt {
